@@ -1,20 +1,22 @@
 name: dart_style
 # Note: See tool/grind.dart for how to bump the version.
-version: 2.3.7-wip
+version: 3.0.1
 description: >-
   Opinionated, automatic Dart source code formatter.
   Provides an API and a CLI tool.
 repository: https://github.com/dart-lang/dart_style
 environment:
-  sdk: "^3.0.0"
+  sdk: "^3.4.0"
 
 dependencies:
-  analyzer: '^6.5.0'
+  analyzer: ">=6.5.0 <8.0.0"
   args: ">=1.0.0 <3.0.0"
   collection: "^1.17.0"
+  package_config: ^2.1.0
   path: ^1.0.0
   pub_semver: ">=1.4.4 <3.0.0"
   source_span: ^1.4.0
+  yaml: ^3.1.2
 
 dev_dependencies:
   dart_flutter_team_lints: ^3.1.0
@@ -22,8 +24,3 @@ dev_dependencies:
   test: ^1.24.6
   test_descriptor: ^2.0.0
   test_process: ^2.0.0
-  yaml: ">=2.0.0 <4.0.0"
-
-executables:
-  dartfmt: format
-  dartformat: format # Allow the old name for compatibility.
