@@ -18,7 +18,7 @@ import 'piece_factory.dart';
 /// delimiter token. Then call [add()] for each [AstNode] that is inside the
 /// delimiters. The [rightBracket()] with the closing delimiter and finally
 /// [build()] to get the resulting [ListPiece].
-class DelimitedListBuilder {
+final class DelimitedListBuilder {
   final PieceFactory _visitor;
 
   /// The opening bracket before the elements, if any.
@@ -51,7 +51,7 @@ class DelimitedListBuilder {
 
   /// Creates the final [ListPiece] out of the added brackets, delimiters,
   /// elements, and style.
-  Piece build() {
+  Piece build({bool forceSplit = false}) {
     // To simplify the piece tree, if there are no elements, just return the
     // brackets concatenated together. We don't have to worry about comments
     // here since they would be in the [_elements] list if there were any.
@@ -62,11 +62,9 @@ class DelimitedListBuilder {
       });
     }
 
-    if (_style.allowBlockElement) _setBlockElementFormatting();
-
     var piece =
         ListPiece(_leftBracket, _elements, _blanksAfter, _rightBracket, _style);
-    if (_mustSplit) piece.pin(State.split);
+    if (_mustSplit || forceSplit) piece.pin(State.split);
     return piece;
   }
 
@@ -134,10 +132,29 @@ class DelimitedListBuilder {
   /// [addCommentsBefore()] for the first token in the [piece].
   ///
   /// Assumes there is no comma after this piece.
-  void add(Piece piece, [BlockFormat format = BlockFormat.none]) {
-    _elements.add(ListElementPiece(_leadingComments, piece, format));
+  void add(Piece piece) {
+    _elements.add(ListElementPiece(_leadingComments, piece));
     _leadingComments.clear();
     _commentsBeforeComma = CommentSequence.empty;
+  }
+
+  /// Adds all of [pieces] to the built list.
+  void addAll(Iterable<Piece> pieces) {
+    pieces.forEach(add);
+  }
+
+  /// Adds the contents of [inner] to this outer [DelimitedListBuilder].
+  ///
+  /// This is used when a [DelimiterListBuilder] is building a piece that will
+  /// then become an element in a surrounding [DelimitedListBuilder]. It ensures
+  /// that any comments around a trailing comma after [inner] don't get lost and
+  /// are instead hoisted up to be captured by this builder.
+  void addInnerBuilder(DelimitedListBuilder inner) {
+    // Add the elements of the line to this builder.
+    add(inner.build());
+
+    // Make sure that any trailing comments on the line aren't lost.
+    _commentsBeforeComma = inner._commentsBeforeComma;
   }
 
   /// Writes any comments appearing before [token] to the list.
@@ -152,23 +169,31 @@ class DelimitedListBuilder {
     // Handle comments between the preceding element and this one.
     addCommentsBefore(element.firstNonCommentToken);
 
-    // See if it's an expression that supports block formatting.
-    var format = switch (element) {
-      AdjacentStrings(indentStrings: true) =>
-        BlockFormat.indentedAdjacentStrings,
-      AdjacentStrings() => BlockFormat.unindentedAdjacentStrings,
-      Expression() => element.blockFormatType,
-      DartPattern() when element.canBlockSplit => BlockFormat.collection,
-      _ => BlockFormat.none,
-    };
-
     // Traverse the element itself.
-    add(_visitor.nodePiece(element), format);
+    add(_visitor.nodePiece(element));
 
     var nextToken = element.endToken.next!;
     if (nextToken.lexeme == ',') {
       _commentsBeforeComma = _visitor.comments.takeCommentsBefore(nextToken);
     }
+  }
+
+  /// Visits a list of [elements].
+  ///
+  /// If [allowBlockArgument] is `true`, then allows one element to receive
+  /// block formatting if appropriate, as in:
+  ///
+  ///     function(argument, [
+  ///       block,
+  ///       like,
+  ///     ], argument);
+  void visitAll(List<AstNode> elements, {bool allowBlockArgument = false}) {
+    for (var i = 0; i < elements.length; i++) {
+      var element = elements[i];
+      visit(element);
+    }
+
+    if (allowBlockArgument) _setBlockArgument(elements);
   }
 
   /// Inserts an inner left delimiter between two elements.
@@ -398,6 +423,14 @@ class DelimitedListBuilder {
     );
   }
 
+  /// Given an argument list, determines which if any of the arguments should
+  /// get special block-like formatting as in the list literal in:
+  ///
+  ///     function(argument, [
+  ///       block,
+  ///       like,
+  ///     ], argument);
+  ///
   /// Looks at the [BlockFormat] types of all of the elements to determine if
   /// one of them should be block formatted.
   ///
@@ -426,85 +459,98 @@ class DelimitedListBuilder {
   ///
   /// Stores the result of this calculation by setting flags on the
   /// [ListElement]s.
-  void _setBlockElementFormatting() {
-    // TODO(tall): These heuristics will probably need some iteration.
-    var functions = <int>[];
-    var collections = <int>[];
-    var adjacentStrings = <int>[];
+  void _setBlockArgument(List<AstNode> arguments) {
+    var candidateIndex = _candidateBlockArgument(arguments);
+    if (candidateIndex == -1) return;
 
-    for (var i = 0; i < _elements.length; i++) {
-      switch (_elements[i].blockFormat) {
-        case BlockFormat.function:
-          functions.add(i);
-        case BlockFormat.collection:
-          collections.add(i);
-        case BlockFormat.invocation:
-          // We don't allow function calls as block elements partially for style
-          // and partially for performance. It often doesn't look great to let
-          // nested function calls pack arbitrarily deeply as block arguments:
-          //
-          //     ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-          //         content: Text(
-          //       localizations.demoSnackbarsAction,
-          //     )));
-          //
-          // This is better when expanded like:
-          //
-          //     ScaffoldMessenger.of(context).showSnackBar(
-          //       SnackBar(
-          //         content: Text(
-          //           localizations.demoSnackbarsAction,
-          //         ),
-          //       ),
-          //     );
-          //
-          // Also, when invocations can be block arguments, which themselves
-          // may contain block arguments, it's easy to run into combinatorial
-          // performance in the solver as it tries to determine which of the
-          // nested calls should and shouldn't be block formatted.
-          break;
-        case BlockFormat.indentedAdjacentStrings:
-        case BlockFormat.unindentedAdjacentStrings:
-          adjacentStrings.add(i);
-        case BlockFormat.none:
-          break; // Not a block element.
-      }
-    }
+    // The block argument must be positional.
+    if (arguments[candidateIndex] is NamedExpression) return;
 
-    switch ((functions, collections, adjacentStrings)) {
-      // Only allow block formatting in an argument list containing adjacent
-      // strings when:
-      //
-      // 1. The block argument is a function expression.
-      // 2. It is the second argument, following an adjacent strings expression.
-      // 3. There are no other adjacent strings in the argument list.
-      //
-      // This matches the `test()` and `group()` and other similar APIs where
-      // you have a message string followed by a block-like function expression
-      // but little else.
-      // TODO(tall): We may want to iterate on these heuristics. For now,
-      // starting with something very narrowly targeted.
-      case ([1], _, [0]):
+    // Only allow up to one trailing argument after the block argument. This
+    // handles the common `tags` and `timeout` named arguments in `test()` and
+    // `group()` while still mostly having the block argument be at the end of
+    // the argument list.
+    if (candidateIndex < arguments.length - 2) return;
+
+    // Edge case: If the first argument is adjacent strings and the second
+    // argument is a function literal, with optionally a third non-block
+    // argument, then treat the function as the block argument.
+    //
+    // This matches the `test()` and `group()` and other similar APIs where
+    // you have a message string followed by a block-like function expression
+    // but little else, as in:
+    //
+    //     test('Some long test description '
+    //         'that splits into multiple lines.', () {
+    //       expect(1 + 2, 3);
+    //     });
+    if (candidateIndex == 1 &&
+        arguments[1].blockFormatType == BlockFormat.function &&
+        arguments[0] is! NamedExpression) {
+      var firstArgumentFormatType = arguments[0].blockFormatType;
+      if (firstArgumentFormatType
+          case BlockFormat.unindentedAdjacentStrings ||
+              BlockFormat.indentedAdjacentStrings) {
         // The adjacent strings.
         _elements[0].allowNewlinesWhenUnsplit = true;
-        if (_elements[0].blockFormat == BlockFormat.unindentedAdjacentStrings) {
+        if (firstArgumentFormatType == BlockFormat.unindentedAdjacentStrings) {
           _elements[0].indentWhenBlockFormatted = true;
         }
 
         // The block-formattable function.
         _elements[1].allowNewlinesWhenUnsplit = true;
-
-      // A function expression takes precedence over other block arguments.
-      case ([var element], _, _):
-        _elements[element].allowNewlinesWhenUnsplit = true;
-
-      // A single collection literal can be block formatted even if there are
-      // other arguments.
-      case ([], [var element], _):
-        _elements[element].allowNewlinesWhenUnsplit = true;
+        return;
+      }
     }
 
-    // If we get here, there are no block element, or it's ambiguous as to
-    // which one should be it so none are.
+    // If we get here, we have a block argument.
+    _elements[candidateIndex].allowNewlinesWhenUnsplit = true;
+  }
+
+  /// If an argument in [arguments] is a candidate to be block formatted,
+  /// returns its index.
+  ///
+  /// If there is a single non-empty block bodied function expression in
+  /// [arguments], returns its index. Otherwise, if there is a single non-empty
+  /// collection literal in [arguments], returns its index. Otherwise, returns
+  /// `-1`.
+  int _candidateBlockArgument(List<AstNode> arguments) {
+    // The index of the function expression argument, or -1 if none has been
+    // found or -2 if there are multiple.
+    var functionIndex = -1;
+
+    // The index of the collection literal argument, or -1 if none has been
+    // found or -2 if there are multiple.
+    var collectionIndex = -1;
+
+    for (var i = 0; i < arguments.length; i++) {
+      // See if it's an expression that supports block formatting.
+      switch (arguments[i].blockFormatType) {
+        case BlockFormat.function:
+          if (functionIndex >= 0) {
+            functionIndex = -2;
+          } else {
+            functionIndex = i;
+          }
+
+        case BlockFormat.collection:
+          if (collectionIndex >= 0) {
+            collectionIndex = -2;
+          } else {
+            collectionIndex = i;
+          }
+
+        case BlockFormat.invocation:
+        case BlockFormat.indentedAdjacentStrings:
+        case BlockFormat.unindentedAdjacentStrings:
+        case BlockFormat.none:
+          break; // Normal argument.
+      }
+    }
+
+    if (functionIndex >= 0) return functionIndex;
+    if (collectionIndex >= 0) return collectionIndex;
+
+    return -1;
   }
 }
