@@ -10,6 +10,7 @@ import '../back_end/solver.dart';
 import '../dart_formatter.dart';
 import '../debug.dart' as debug;
 import '../piece/adjacent.dart';
+import '../piece/leading_comment.dart';
 import '../piece/list.dart';
 import '../piece/piece.dart';
 import '../piece/text.dart';
@@ -24,7 +25,7 @@ import 'sequence_builder.dart';
 ///
 /// Handles updating selection markers and attaching comments to the tokens
 /// before and after the comments.
-class PieceWriter {
+final class PieceWriter {
   final DartFormatter _formatter;
 
   final SourceCode _source;
@@ -96,11 +97,11 @@ class PieceWriter {
 
     if (spaceBefore) space();
 
-    // TODO(perf): If [_currentCode] is `null` but [_pendingSpace] is `true`,
-    // it should be possible to create a new code piece and write the leading
-    // space to it instead of having a leading SpacePiece. Unfortunately, that
-    // sometimes leads to duplicate spaces in the output, so it might take some
-    // tweaking to get working.
+    // TODO(rnystrom): If [_currentCode] is `null` but [_pendingSpace] is
+    // `true`, it should be possible to create a new code piece and write the
+    // leading space to it instead of having a leading SpacePiece.
+    // Unfortunately, that sometimes leads to duplicate spaces in the output,
+    // so it might take some tweaking to get working.
 
     if (token.precedingComments != null) {
       // Don't append to the previous token if there is a comment after it.
@@ -177,11 +178,18 @@ class PieceWriter {
     _flushSpace();
     _currentCode = null;
 
-    var metadataPieces = const <Piece>[];
+    var leadingPieces = const <Piece>[];
     if (metadata.isNotEmpty) {
-      metadataPieces = [
+      leadingPieces = [
         for (var annotation in metadata) _visitor.nodePiece(annotation)
       ];
+
+      // If there are comments between the metadata and declaration, then hoist
+      // them out too so they don't get embedded inside the beginning piece of
+      // the declaration. [SequenceBuilder] handles that for most comments
+      // preceding a declaration but won't see these ones because they come
+      // after the metadata.
+      leadingPieces.addAll(takeCommentsBefore(metadata.last.endToken.next!));
     }
 
     _pieces.add([]);
@@ -198,7 +206,7 @@ class PieceWriter {
         ? builtPieces.first
         : AdjacentPiece(builtPieces);
 
-    if (metadataPieces.isEmpty) {
+    if (leadingPieces.isEmpty) {
       // No metadata, so return the content piece directly.
       return builtPiece;
     } else if (inlineMetadata) {
@@ -210,7 +218,7 @@ class PieceWriter {
             spaceWhenUnsplit: true,
           ));
 
-      for (var piece in metadataPieces) {
+      for (var piece in leadingPieces) {
         list.add(piece);
       }
 
@@ -219,7 +227,7 @@ class PieceWriter {
     } else {
       // Wrap the metadata and content in a sequence.
       var sequence = SequenceBuilder(_visitor);
-      for (var piece in metadataPieces) {
+      for (var piece in leadingPieces) {
         sequence.add(piece);
       }
 
@@ -269,15 +277,21 @@ class PieceWriter {
     }
   }
 
-  // TODO(tall): Much of the comment handling code in CommentWriter got moved
-  // into here, so there isn't great separation of concerns anymore. Can we
-  // organize this code better? Or just combine CommentWriter with this class
-  // completely?
-
   /// Creates a new [Piece] for [comment] and returns it.
   Piece commentPiece(SourceComment comment,
       [Whitespace trailingWhitespace = Whitespace.none]) {
-    var piece = CommentPiece(trailingWhitespace);
+    var piece = switch (comment.text) {
+      '// dart format off' => EnableFormattingCommentPiece(
+          enable: false,
+          comment.offset + comment.text.length,
+          trailingWhitespace),
+      '// dart format on' => EnableFormattingCommentPiece(
+          enable: true,
+          comment.offset + comment.text.length,
+          trailingWhitespace),
+      _ => CommentPiece(trailingWhitespace),
+    };
+
     _write(piece, comment.text, comment.offset,
         multiline: comment.type.mayBeMultiline);
     return piece;
@@ -287,6 +301,28 @@ class PieceWriter {
   /// and takes and returns any remaining leading comments.
   List<Piece> takeCommentsBefore(Token token) {
     return _splitComments(_comments.takeCommentsBefore(token), token);
+  }
+
+  /// Takes any comments preceding [firstToken] and wraps them around the piece
+  /// generated by [buildCallback].
+  ///
+  /// For comments preceding a single AST node, this hoisting is handled
+  /// automatically. But there are some places in the language where there is
+  /// a conceptual piece of syntax that we want to hoist the comments out of
+  /// so that the syntax doesn't split but there isn't actually a single AST
+  /// node associated with that syntax.
+  ///
+  /// This method lets you hoist comments before an arbitary amount of syntax
+  /// visited and built by calling [buildCallback].
+  void hoistLeadingComments(Token firstToken, Piece Function() buildCallback) {
+    var leadingComments = takeCommentsBefore(firstToken);
+
+    var piece = buildCallback();
+    if (leadingComments.isNotEmpty) {
+      piece = LeadingCommentPiece(leadingComments, piece);
+    }
+
+    add(piece);
   }
 
   /// Begins a new [CodeToken] that can potentially have more code written to
@@ -379,7 +415,11 @@ class PieceWriter {
 
   /// Finishes writing and returns a [SourceCode] containing the final output
   /// and updated selection, if any.
-  SourceCode finish(Piece rootPiece) {
+  ///
+  /// If there is a `// dart format width=123` comment before the formatted
+  /// code, then [pageWidthFromComment] is that width.
+  SourceCode finish(
+      SourceCode source, Piece rootPiece, int? pageWidthFromComment) {
     if (debug.tracePieceBuilder) {
       debug.log(debug.pieceTree(rootPiece));
     }
@@ -387,35 +427,15 @@ class PieceWriter {
     Profile.begin('PieceWriter.finish() format piece tree');
 
     var cache = SolutionCache();
-    var formatter = Solver(cache,
-        pageWidth: _formatter.pageWidth, leadingIndent: _formatter.indent);
-    var result = formatter.format(rootPiece);
-    var outputCode = result.text;
+    var solver = Solver(cache,
+        pageWidth: pageWidthFromComment ?? _formatter.pageWidth,
+        leadingIndent: _formatter.indent);
+    var solution = solver.format(rootPiece);
+    var output = solution.code.build(source, _formatter.lineEnding);
 
     Profile.end('PieceWriter.finish() format piece tree');
 
-    // Be a good citizen, end with a newline.
-    if (_source.isCompilationUnit) outputCode += _formatter.lineEnding!;
-
-    int? selectionStart;
-    int? selectionLength;
-    if (_source.selectionStart != null) {
-      selectionStart = result.selectionStart;
-      var selectionEnd = result.selectionEnd;
-
-      // If we haven't hit the beginning and/or end of the selection yet, they
-      // must be at the very end of the code.
-      selectionStart ??= outputCode.length;
-      selectionEnd ??= outputCode.length;
-
-      selectionLength = selectionEnd - selectionStart;
-    }
-
-    return SourceCode(outputCode,
-        uri: _source.uri,
-        isCompilationUnit: _source.isCompilationUnit,
-        selectionStart: selectionStart,
-        selectionLength: selectionLength);
+    return output;
   }
 
   /// Returns the number of characters past [position] in the source where the
