@@ -7,14 +7,21 @@ import '../constants.dart';
 import 'piece.dart';
 
 /// Piece for a case pattern, guard, and body in a switch expression.
-class CaseExpressionPiece extends Piece {
+final class CaseExpressionPiece extends Piece {
+  /// Split inside the body, which must be block formattable, like:
+  ///
+  ///     pattern => function(
+  ///       argument,
+  ///     ),
+  static const State _blockSplitBody = State(1, cost: 0);
+
   /// Split after the `=>` before the body.
-  static const State _beforeBody = State(1);
+  static const State _beforeBody = State(2);
 
   /// Split before the `when` guard clause and after the `=>`.
-  static const State _beforeWhenAndBody = State(2);
+  static const State _beforeWhenAndBody = State(3);
 
-  /// The pattern the value is matched against along with the leading `case`.
+  /// The pattern the value is matched against.
   final Piece _pattern;
 
   /// If there is a `when` clause, that clause.
@@ -54,6 +61,7 @@ class CaseExpressionPiece extends Piece {
 
   @override
   List<State> get additionalStates => [
+        if (_canBlockSplitBody) _blockSplitBody,
         _beforeBody,
         if (_guard != null) ...[_beforeWhenAndBody],
       ];
@@ -61,10 +69,15 @@ class CaseExpressionPiece extends Piece {
   @override
   bool allowNewlineInChild(State state, Piece child) {
     return switch (state) {
+      // If the outermost pattern is `||`, then always let it split even while
+      // allowing the body on the same line as `=>`.
+      _ when child == _pattern && _patternIsLogicalOr => true,
+
+      // There are almost never splits in the arrow piece. It requires a comment
+      // in a funny location, but if it happens, allow it.
       _ when child == _arrow => true,
-      State.unsplit when child == _body && _canBlockSplitBody => true,
-      _beforeBody when child == _pattern =>
-        _guard == null || _patternIsLogicalOr,
+      _blockSplitBody when child == _body => true,
+      _beforeBody when child == _pattern => _guard == null,
       _beforeBody when child == _body => true,
       _beforeWhenAndBody => true,
       _ => false,
@@ -94,12 +107,13 @@ class CaseExpressionPiece extends Piece {
     writer.space();
     writer.format(_arrow);
 
-    if (state != State.unsplit) writer.pushIndent(Indent.block);
+    var indentBody = state != State.unsplit && state != _blockSplitBody;
+    if (indentBody) writer.pushIndent(Indent.block);
 
     writer.splitIf(state == _beforeBody || state == _beforeWhenAndBody);
     writer.format(_body);
 
-    if (state != State.unsplit) writer.popIndent();
+    if (indentBody) writer.popIndent();
   }
 
   @override
