@@ -8,28 +8,7 @@ import 'piece.dart';
 /// A piece for a series of binary expressions at the same precedence, like:
 ///
 ///     a + b + c
-class InfixPiece extends Piece {
-  /// Pieces for leading comments that appear before the first operand.
-  ///
-  /// We hoist these comments out from the first operand's first token so that
-  /// a newline in these comments doesn't erroneously force the infix operator
-  /// to split. For example:
-  ///
-  ///     value =
-  ///         // comment
-  ///         a + b;
-  ///
-  /// Here, the `// comment` will be hoisted out and stored in
-  /// [_leadingComments] instead of being a leading comment in the [CodePiece]
-  /// for `a`. If we left the comment in `a`, then the newline after the line
-  /// comment would force the `+` operator to split yielding:
-  ///
-  ///     value =
-  ///         // comment
-  ///         a +
-  ///             b;
-  final List<Piece> _leadingComments;
-
+final class InfixPiece extends Piece {
   /// The series of operands.
   ///
   /// Since we don't split on both sides of the operator, the operators will be
@@ -41,26 +20,16 @@ class InfixPiece extends Piece {
   /// Whether operands after the first should be indented if split.
   final bool _indent;
 
-  InfixPiece(this._leadingComments, this._operands, {bool indent = true})
-      : _indent = indent;
+  InfixPiece(this._operands, {bool indent = true}) : _indent = indent;
 
   @override
   List<State> get additionalStates => const [State.split];
 
   @override
-  bool allowNewlineInChild(State state, Piece child) {
-    if (state == State.split) return true;
-
-    // Comments before the operands don't force the operator to split.
-    return _leadingComments.contains(child);
-  }
+  bool allowNewlineInChild(State state, Piece child) => state == State.split;
 
   @override
   void format(CodeWriter writer, State state) {
-    for (var comment in _leadingComments) {
-      writer.format(comment);
-    }
-
     if (_indent) writer.pushIndent(Indent.expression);
 
     for (var i = 0; i < _operands.length; i++) {
@@ -78,7 +47,6 @@ class InfixPiece extends Piece {
 
   @override
   void forEachChild(void Function(Piece piece) callback) {
-    _leadingComments.forEach(callback);
     _operands.forEach(callback);
   }
 
