@@ -38,7 +38,7 @@ import 'piece.dart';
 ///
 /// ListPieces are usually constructed using [createList()] or
 /// [DelimitedListBuilder].
-class ListPiece extends Piece {
+final class ListPiece extends Piece {
   /// The opening bracket before the elements, if any.
   final Piece? _before;
 
@@ -252,9 +252,6 @@ final class ListElementPiece extends Piece {
 
   final Piece? _content;
 
-  /// What kind of block formatting can be applied to this element.
-  final BlockFormat blockFormat;
-
   /// Whether newlines are allowed in this element when this list is unsplit.
   ///
   /// This is generally only true for a single "block" element, as in:
@@ -312,16 +309,13 @@ final class ListElementPiece extends Piece {
   /// delimiter (here `,` and 2).
   int _commentsBeforeDelimiter = 0;
 
-  ListElementPiece(
-      List<Piece> leadingComments, Piece element, BlockFormat format)
+  ListElementPiece(List<Piece> leadingComments, Piece element)
       : _leadingComments = [...leadingComments],
-        _content = element,
-        blockFormat = format;
+        _content = element;
 
   ListElementPiece.comment(Piece comment)
       : _leadingComments = const [],
-        _content = null,
-        blockFormat = BlockFormat.none {
+        _content = null {
     _hangingComments.add(comment);
   }
 
@@ -404,9 +398,10 @@ enum BlockFormat {
   /// elements.
   function,
 
-  /// The element is a collection literal.
+  /// The element is a collection literal or multiline string literal.
   ///
-  /// These can be block formatted even when there are other arguments.
+  /// If there is only one of these and no [BlockFormat.function] elements, then
+  /// it can be block formatted.
   collection,
 
   /// A function or method invocation.
@@ -416,7 +411,7 @@ enum BlockFormat {
 
   /// The element is an adjacent strings expression that's in an list that
   /// requires its subsequent lines to be indented (because there are other
-  /// string literal in the list).
+  /// string literals in the list).
   indentedAdjacentStrings,
 
   /// The element is an adjacent strings expression that's in an list that
@@ -440,7 +435,7 @@ enum BlockFormat {
 /// they vary in whether or not a trailing comma is allowed, whether there
 /// should be spaces inside the delimiters when the elements aren't split, etc.
 /// This class captures those options.
-class ListStyle {
+final class ListStyle {
   /// How commas should be handled by the list.
   ///
   /// Most lists use [Commas.trailing]. Type parameters and type arguments use
@@ -459,18 +454,8 @@ class ListStyle {
   ///     //              ^                      ^
   final bool spaceWhenUnsplit;
 
-  /// Whether an element in the list is allowed to have block-like formatting,
-  /// as in:
-  ///
-  ///     function(argument, [
-  ///       block,
-  ///       like,
-  ///     ], argument);
-  final bool allowBlockElement;
-
   const ListStyle(
       {this.commas = Commas.trailing,
       this.splitCost = Cost.normal,
-      this.spaceWhenUnsplit = false,
-      this.allowBlockElement = false});
+      this.spaceWhenUnsplit = false});
 }
