@@ -2,10 +2,10 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:dart_style/dart_style.dart';
-import 'package:dart_style/src/constants.dart';
 import 'package:dart_style/src/testing/benchmark.dart';
 import 'package:dart_style/src/testing/test_file.dart';
 import 'package:path/path.dart' as p;
@@ -20,124 +20,72 @@ const formattedSource = 'void main() => print("hello");\n';
 /// [TestProcess] filters those when it strips command line output into lines.
 const formattedOutput = 'void main() => print("hello");';
 
-/// If tool/command_shell.dart has been compiled to a snapshot, this is the path
-/// to it.
-String? _commandExecutablePath;
+/// If `bin/format.dart` has been compiled to a snapshot, this is the path to
+/// it.
+String? _formatterPath;
 
-/// If bin/format.dart has been compiled to a snapshot, this is the path to it.
-String? _formatterExecutablePath;
-
-/// Compiles format.dart to a native executable for tests to use.
+/// Compiles `bin/format.dart` to a native executable for tests to use.
 ///
-/// Calls [setupAll()] and [tearDownAll()] to coordinate this when the
+/// Calls [setupAll()] and [tearDownAll()] to coordinate this with the
 /// subsequent tests and to clean up the executable.
-void compileFormatterExecutable() {
+void compileFormatter() {
   setUpAll(() async {
-    _formatterExecutablePath = await _compileSnapshot('bin/format.dart');
+    var tempDir =
+        await Directory.systemTemp.createTemp(p.withoutExtension('format'));
+    _formatterPath = p.join(tempDir.path, 'format.dart.snapshot');
+    var scriptPath =
+        p.normalize(p.join(await findTestDirectory(), '../bin/format.dart'));
+
+    var compileResult = await Process.run(Platform.resolvedExecutable, [
+      '--snapshot-kind=app-jit',
+      '--snapshot=$_formatterPath',
+      scriptPath,
+      '--help'
+    ]);
+
+    if (compileResult.exitCode != 0) {
+      fail('Could not compile format.dart to a snapshot (exit code '
+          '${compileResult.exitCode}):\n${compileResult.stdout}\n\n'
+          '${compileResult.stderr}');
+    }
   });
 
   tearDownAll(() async {
-    await _deleteSnapshot(_formatterExecutablePath!);
-    _formatterExecutablePath = null;
+    try {
+      await Directory(p.dirname(_formatterPath!)).delete(recursive: true);
+    } on IOException {
+      // Do nothing if we failed to delete it. The OS will eventually clean it
+      // up.
+    }
   });
 }
 
-/// Compiles command_shell.dart to a native executable for tests to use.
-///
-/// Calls [setupAll()] and [tearDownAll()] to coordinate this when the
-/// subsequent tests and to clean up the executable.
-void compileCommandExecutable() {
-  setUpAll(() async {
-    _commandExecutablePath = await _compileSnapshot('tool/command_shell.dart');
-  });
-
-  tearDownAll(() async {
-    await _deleteSnapshot(_commandExecutablePath!);
-    _commandExecutablePath = null;
-  });
-}
-
-/// Compile the Dart [script] to an app-JIT snapshot.
-///
-/// We do this instead of spawning the script from source each time because it's
-/// much faster when the same script needs to be run several times.
-Future<String> _compileSnapshot(String script) async {
-  var scriptName = p.basename(script);
-  var tempDir =
-      await Directory.systemTemp.createTemp(p.withoutExtension(scriptName));
-  var snapshot = p.join(tempDir.path, '$scriptName.snapshot');
-  var scriptPath = p.normalize(p.join(await findTestDirectory(), '..', script));
-
-  var compileResult = await Process.run(Platform.resolvedExecutable, [
-    '--snapshot-kind=app-jit',
-    '--snapshot=$snapshot',
-    scriptPath,
-    '--help'
-  ]);
-
-  if (compileResult.exitCode != 0) {
-    fail('Could not compile $scriptName to a snapshot (exit code '
-        '${compileResult.exitCode}):\n${compileResult.stdout}\n\n'
-        '${compileResult.stderr}');
-  }
-
-  return snapshot;
-}
-
-/// Attempts to delete to temporary directory created for [snapshot] by
-/// [_compileSnapshot()].
-Future<void> _deleteSnapshot(String snapshot) async {
-  try {
-    await Directory(p.dirname(snapshot)).delete(recursive: true);
-  } on IOException {
-    // Do nothing if we failed to delete it. The OS will eventually clean it
-    // up.
-  }
-}
-
-/// Runs the command line formatter, passing it [args].
+/// Runs the command-line formatter, passing it [args].
 Future<TestProcess> runFormatter([List<String>? args]) {
-  if (_formatterExecutablePath == null) {
-    fail('Must call createFormatterExecutable() before running commands.');
+  if (_formatterPath == null) {
+    fail('Must call createCommandExecutable() before running commands.');
   }
 
   return TestProcess.start(
-      Platform.resolvedExecutable, [_formatterExecutablePath!, ...?args],
+      Platform.resolvedExecutable, [_formatterPath!, ...?args],
       workingDirectory: d.sandbox);
 }
 
-/// Runs the command line formatter, passing it the test directory followed by
+/// Runs the command-line formatter, passing it the test directory followed by
 /// [args].
 Future<TestProcess> runFormatterOnDir([List<String>? args]) {
   return runFormatter(['.', ...?args]);
 }
 
-/// Runs the test shell for the [Command]-based formatter, passing it [args].
-Future<TestProcess> runCommand([List<String>? args]) {
-  if (_commandExecutablePath == null) {
-    fail('Must call createCommandExecutable() before running commands.');
-  }
-
-  return TestProcess.start(Platform.resolvedExecutable,
-      [_commandExecutablePath!, 'format', ...?args],
-      workingDirectory: d.sandbox);
-}
-
-/// Runs the test shell for the [Command]-based formatter, passing it the test
-/// directory followed by [args].
-Future<TestProcess> runCommandOnDir([List<String>? args]) {
-  return runCommand(['.', ...?args]);
-}
-
 /// Run tests defined in "*.unit" and "*.stmt" files inside directory [path].
-Future<void> testDirectory(String path, {Iterable<StyleFix>? fixes}) async {
+Future<void> testDirectory(String path) async {
   for (var test in await TestFile.listDirectory(path)) {
-    _testFile(test, fixes);
+    _testFile(test);
   }
 }
 
-Future<void> testFile(String path, {Iterable<StyleFix>? fixes}) async {
-  _testFile(await TestFile.read(path), fixes);
+Future<void> testFile(String path) async {
+  _testFile(await TestFile.read(path));
 }
 
 /// Format all of the benchmarks and ensure they produce their expected outputs.
@@ -148,10 +96,11 @@ Future<void> testBenchmarks({required bool useTallStyle}) async {
     for (var benchmark in benchmarks) {
       test(benchmark.name, () {
         var formatter = DartFormatter(
+            languageVersion: useTallStyle
+                ? DartFormatter.latestLanguageVersion
+                : DartFormatter.latestShortStyleLanguageVersion,
             pageWidth: benchmark.pageWidth,
-            experimentFlags: useTallStyle
-                ? const ['inline-class', 'macros', tallStyleExperimentFlag]
-                : const ['inline-class', 'macros']);
+            experimentFlags: const ['macros']);
 
         var actual = formatter.formatSource(SourceCode(benchmark.input));
 
@@ -175,50 +124,128 @@ Future<void> testBenchmarks({required bool useTallStyle}) async {
   });
 }
 
-void _testFile(TestFile testFile, Iterable<StyleFix>? baseFixes) {
-  var useTallStyle = testFile.path.startsWith('tall/');
-
+void _testFile(TestFile testFile) {
   group(testFile.path, () {
     for (var formatTest in testFile.tests) {
       test(formatTest.label, () {
-        var fixes = [...?baseFixes, ...formatTest.fixes];
-
-        if (useTallStyle && fixes.isNotEmpty) {
-          fail('Test error: Tall style does not support applying fixes.');
-        }
-
         var formatter = DartFormatter(
+            languageVersion: formatTest.languageVersion,
             pageWidth: testFile.pageWidth,
             indent: formatTest.leadingIndent,
-            fixes: fixes,
-            experimentFlags: useTallStyle
-                ? const ['inline-class', 'macros', tallStyleExperimentFlag]
-                : const ['inline-class', 'macros']);
+            experimentFlags: const ['macros']);
 
-        var actual = formatter.formatSource(formatTest.input);
+        var actual = _validateFormat(
+            formatter,
+            formatTest.input,
+            formatTest.output,
+            'did not match expectation',
+            testFile.isCompilationUnit);
 
-        // The test files always put a newline at the end of the expectation.
-        // Statements from the formatter (correctly) don't have that, so add
-        // one to line up with the expected result.
-        var actualText = actual.text;
-        if (!testFile.isCompilationUnit) actualText += '\n';
-
-        // Fail with an explicit message because it's easier to read than
-        // the matcher output.
-        if (actualText != formatTest.output.text) {
-          fail('Formatting did not match expectation. Expected:\n'
-              '${formatTest.output.text}\nActual:\n$actualText');
-        } else if (actual.selectionStart != formatTest.output.selectionStart ||
-            actual.selectionLength != formatTest.output.selectionLength) {
-          fail('Selection did not match expectation. Expected:\n'
-              '${formatTest.output.textWithSelectionMarkers}\n'
-              'Actual:\n${actual.textWithSelectionMarkers}');
-        }
-
-        expect(actual.selectionStart, equals(formatTest.output.selectionStart));
-        expect(
-            actual.selectionLength, equals(formatTest.output.selectionLength));
+        // Make sure that formatting is idempotent. Format the output and make
+        // sure we get the same result.
+        _validateFormat(formatter, actual, actual, 'was not idempotent',
+            testFile.isCompilationUnit);
       });
     }
   });
+}
+
+/// Run [formatter] on [input] and validate that the result matches [expected].
+///
+/// If not, fails with an error using [reason].
+///
+/// Returns the formatted output.
+SourceCode _validateFormat(DartFormatter formatter, SourceCode input,
+    SourceCode expected, String reason, bool isCompilationUnit) {
+  var actual = formatter.formatSource(input);
+
+  // Fail with an explicit message because it's easier to read than
+  // the matcher output.
+  if (actual.text != expected.text) {
+    fail('Formatting $reason. Expected:\n'
+        '${expected.text}\nActual:\n${actual.text}');
+  } else if (actual.selectionStart != expected.selectionStart ||
+      actual.selectionLength != expected.selectionLength) {
+    fail('Selection $reason. Expected:\n'
+        '${expected.textWithSelectionMarkers}\n'
+        'Actual:\n${actual.textWithSelectionMarkers}');
+  }
+
+  return actual;
+}
+
+/// Create a test `.dart_tool` directory with a package config for a package
+/// with [rootPackageName] and language version [major].[minor].
+///
+/// If [packages] is given, it should be a map from package names to root URIs
+/// for each package.
+d.DirectoryDescriptor packageConfig(String rootPackageName,
+    {String? version, Map<String, String>? packages}) {
+  var defaultVersion = DartFormatter.latestLanguageVersion;
+  version ??= '${defaultVersion.major}.${defaultVersion.minor}';
+
+  Map<String, dynamic> package(String name, String rootUri) => {
+        'name': name,
+        'rootUri': rootUri,
+        'packageUri': 'lib/',
+        'languageVersion': version
+      };
+
+  var config = {
+    'configVersion': 2,
+    'packages': [
+      package(rootPackageName, '../'),
+      if (packages != null)
+        for (var name in packages.keys) package(name, packages[name]!),
+    ]
+  };
+
+  return d.dir('.dart_tool', [
+    d.file('package_config.json', jsonEncode(config)),
+  ]);
+}
+
+/// Creates the YAML string contents of an analysis options file.
+///
+/// If [pageWidth] is given, then the result has a "formatter" section to
+/// specify the page width. If [include] is given, then adds an "include" key
+/// to include another analysis options file. If [other] is given, then those
+/// are added as other top-level keys in the YAML.
+String analysisOptions(
+    {int? pageWidth,
+    Object? /* String | List<String> */ include,
+    Map<String, Object>? other}) {
+  var yaml = StringBuffer();
+
+  switch (include) {
+    case String _:
+      yaml.writeln('include: $include');
+    case List<String> _:
+      yaml.writeln('include:');
+      for (var path in include) {
+        yaml.writeln('  - $path');
+      }
+  }
+
+  if (pageWidth != null) {
+    yaml.writeln('formatter:');
+    yaml.writeln('  page_width: $pageWidth');
+  }
+
+  if (other != null) {
+    other.forEach((key, value) {
+      yaml.writeln('$key:');
+      yaml.writeln('  $value');
+    });
+  }
+
+  return yaml.toString();
+}
+
+/// Creates a file named "analysis_options.yaml" containing the given YAML
+/// options to configure the [pageWidth] and [include] file, if any.
+d.FileDescriptor analysisOptionsFile(
+    {String name = 'analysis_options.yaml', int? pageWidth, String? include}) {
+  var yaml = analysisOptions(pageWidth: pageWidth, include: include);
+  return d.FileDescriptor(name, yaml.toString());
 }
