@@ -4,7 +4,6 @@
 import 'dart:io';
 
 import 'package:dart_style/dart_style.dart';
-import 'package:dart_style/src/constants.dart';
 import 'package:dart_style/src/testing/test_file.dart';
 import 'package:path/path.dart' as p;
 
@@ -87,37 +86,14 @@ Future<void> _updateTestFile(TestFile testFile) async {
   // Write the file-level comments.
   _writeComments(buffer, testFile.comments);
 
-  // TODO(rnystrom): This is duplicating logic in fix_test.dart. Ideally, we'd
-  // move the fix markers into the tests themselves, but since --fix is
-  // probably going away, it's not worth it.
-  var baseFixes = const {
-        'short/fixes/doc_comments.stmt': [StyleFix.docComments],
-        'short/fixes/function_typedefs.unit': [StyleFix.functionTypedefs],
-        'short/fixes/named_default_separator.unit': [
-          StyleFix.namedDefaultSeparator
-        ],
-        'short/fixes/optional_const.unit': [StyleFix.optionalConst],
-        'short/fixes/optional_new.stmt': [StyleFix.optionalNew],
-        'short/fixes/single_cascade_statements.stmt': [
-          StyleFix.singleCascadeStatements
-        ],
-      }[testFile.path] ??
-      const <StyleFix>[];
-
-  var experiments = [
-    'inline-class',
-    'macros',
-    if (p.split(testFile.path).contains('tall')) tallStyleExperimentFlag
-  ];
-
   _totalTests += testFile.tests.length;
 
   for (var formatTest in testFile.tests) {
     var formatter = DartFormatter(
+        languageVersion: formatTest.languageVersion,
         pageWidth: testFile.pageWidth,
         indent: formatTest.leadingIndent,
-        fixes: [...baseFixes, ...formatTest.fixes],
-        experimentFlags: experiments);
+        experimentFlags: const ['macros']);
 
     var actual = formatter.formatSource(formatTest.input);
 
@@ -130,9 +106,15 @@ Future<void> _updateTestFile(TestFile testFile) async {
     // Insert a newline between each test, but not after the last.
     if (formatTest != testFile.tests.first) buffer.writeln();
 
+    var defaultLanguageVersion = p.split(testFile.path).contains('tall')
+        ? DartFormatter.latestLanguageVersion
+        : DartFormatter.latestShortStyleLanguageVersion;
+
     var descriptionParts = [
       if (formatTest.leadingIndent != 0) '(indent ${formatTest.leadingIndent})',
-      for (var fix in formatTest.fixes) '(fix ${fix.name})',
+      if (formatTest.languageVersion != defaultLanguageVersion)
+        '(version ${formatTest.languageVersion.major}.'
+            '${formatTest.languageVersion.minor})',
       formatTest.description
     ];
 
